requirements.txt
cryptography==41.0.3
fastapi~=0.95.2
orjson~=3.8.8
uvicorn
jinja2~=3.1.2
pycryptodome==3.18.0